fix: size_t overflow

+++ src/bubble_sort.cpp
@@ -1,7 +1,7 @@
 #include <bubble_sort.hpp>
 
 void BubbleSort(std::vector<int> &a) {
-  auto n = a.size();
+  auto n = static_cast<int>(a.size());
   for (int i = 0; i < n; i++) {
     for (int j = 0; j < n - i - 1; j++) {
       if (a[j] > a[i]) {
